auth_service/requirements.txt
fastapi
uvicorn
pymongo
argon2-cffi
python-dotenv
pyjwt
